fix: infinite loop on removal of last document in a workflow from its editor
Problem: When deleting the last document of a workflow, the `Workflows.deleteWorkflow` method is
called and closes the corresponding workflow editor. However, by attempting to close the editor, a
save is triggered, calling the `Workflows.saveWorkflowOrDeleteIfEmpty` method which calls the
`Workflows.deleteWorkflow` method again, causing an infinite loop.

Solution: Introduce a `isClosing` field to the `Editor` class set to `true` when the `close` method
of an editor is first called.